requirements.txt
requests==2.32.3
neo4j==5.26.0
pandas==2.2.2
streamlit==1.38.0
openai==1.8.0
python-dotenv==1.0.0
plotly==5.24.0
pyvis==0.3.2
numpy==1.26.4